+++ include/f16/reply.hpp
@@ -75,6 +75,8 @@ struct reply
     unauthorized = 401,
     forbidden = 403,
     not_found = 404,
+    request_timeout = 408,
+    request_entity_too_large = 413,
     internal_server_error = 500,
     not_implemented = 501,
     bad_gateway = 502,